requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
sqlalchemy==2.0.23
torch==2.1.0
transformers==4.35.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2 